Refactor code to handle bucket files instead of images
The commit replaces all usages of "BucketImage" objects with "BucketFile" in the gallery services to make the code more generic. This allows handling of different file types rather than just images. It also changes file extensions from '.jpg' to '.webp' in the 'populateGalleries' function. As part of debugging, a console log was added in the 'populateGalleryImages' function.