Add home page slider image support
This commit incorporates the support for slider images to the home page. It imports additional functions and constants, introduces a new "slider" prop in the HomePage component, and fetches the "home_slider" image in the getStaticProps function. This change enables the application to dynamically render the slider image on the home page.